+++ src/viaje/LargaDistancia.java
@@ -22,7 +22,7 @@ public class LargaDistancia extends Viaje{
 	/*CREAR EXCEPCION EN CASO DE QUE EL DESTINO SEA MENOR A 100 KM*/
 	/*CHEQUEAR QUE EL TRANSPORTE SEA SOLO COLECTIVO O COMBI*/
 	
-	public double Costo(){
+	public float Costo(){
 		Responsable nodo = null;
 		float sum=0;
 		float sumaSueldo = 0;
